fix(vite-config): make the package.json sync a build-only side effect
Running knip in a workspace rewrote that workspace's package.json,
flipping main/module/types/exports from the src entry points to dist
paths. Knip is read-only; the write came from us.

knip's vite plugin loads each workspace's vite.config.ts to discover
entry points. Those configs are `export default createViteConfig(...)`,
and createViteConfig called updatePackageJson() eagerly while building
the config object - so merely *evaluating* the config wrote to disk. No
`vite` script or build was needed, only a tool that reads the config.

Separate the pure derivation from the write, and hang the write off a
build lifecycle hook:

- lib/package-json/generate.ts - generatePackageJsonFields, no fs
- lib/package-json/write.ts    - updatePackageJson, the fs write
- lib/package-json/plugin.ts   - createPackageJsonPlugin, apply: "build"

Constructing a plugin is inert, so config loaders now get a plugin
object and nothing else. The hook is closeBundle rather than buildStart,
so the fields are written only once the artifacts they point at exist -
a failed build no longer leaves the manifest aimed at a dist/ that was
never produced. syncPackageJson() stays as the explicit escape hatch for
callers that want the write without a build.

Verified both directions on packages/ui/nfl: knip leaves package.json
untouched, `vite build` still syncs it.
