+++ packages/some-vite-config/src/lib/index.ts
@@ -2,4 +2,4 @@ export * from "./build-config.js"
 export * from "./plugins.js"
 export * from "./resolve.js"
 export * from "./rollup.js"
-export * from "./package-json.js"
+export * from "./package-json/index.js"
