+++ packages/some-vite-config/src/lib/package-json/index.ts
@@ -1,0 +1,3 @@
+export * from "./generate.js"
+export * from "./plugin.js"
+export * from "./write.js"
